% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Assign number variable to bins}
\usage{
bin_variable(x, bin.width)
}
\arguments{
\item{x}{A vector}

\item{bin.width}{The bin width. E.g., 20 will make each bin a maximum of 20 units wide}
}
\value{
A data frame
}
\description{
Useful before summarizing and plotting.
}
\examples{
x<-c(seq(1,23,1),seq(5,12,1))
bin_variable(x,5)
}
