% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eclectic.R
\docType{data}
\name{blast6_colnames}
\alias{blast6_colnames}
\title{Default BLAST outfmt 6 column names}
\format{\preformatted{ chr [1:12] "qseqid" "sseqid" "pident" "length" "mismatch" ...
}}
\usage{
blast6_colnames
}
\description{
Default BLAST outfmt 6 column names
}
\keyword{datasets}

