% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{contam_test}
\alias{contam_test}
\title{Test for an OTU's (inverse) correlation between its abundance and amplicon concentration.}
\usage{
contam_test(.df, ..., otu = "otu", freq = "freq", conc = "PostPCRConc")
}
\arguments{
\item{.df}{the data frame}

\item{...}{the grouping variables (unquoted), i.e. ExtractionType and SampleType}

\item{otu}{the column with unique otu names}

\item{freq}{the column denoting OTU proportional abundance (as string)}

\item{conc}{the column containing amplicon concentration data}
}
\value{
the data frame with correlation results added in for each OTU
}
\description{
Test for an OTU's (inverse) correlation between its abundance and amplicon concentration.
}

