% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Count occurrences of NA}
\usage{
count_na(x, count.empty = FALSE)
}
\arguments{
\item{x}{A vector or string}

\item{count.empty}{Treat empty values (white space) as NA? Logical. Default is FALSE.}
}
\description{
Count occurrences of NA
}
\examples{
count_na(c(1,2,NA,4,NA))# 2
count_na(c(1,2,'',4,NA))# 1
count_na(c(1,2,'',4,NA),count.empty=TRUE)# 2
count_na(c(1,2,'    ',4,NA),count.empty=TRUE)# 2
}
