% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{dynamic_chunk}
\alias{dynamic_chunk}
\title{Creates a chunk for a plot with a given figure height and width.}
\usage{
dynamic_chunk(p, width, height, name = NULL, extras = NULL)
}
\arguments{
\item{p}{the plot to insert into the chunk}

\item{width}{the width of the device}

\item{height}{the height of the device}

\item{name}{the name of the chunk}

\item{extras}{a string to add to the end of the chunk definition, for kicks}
}
\description{
Creates a chunk for a plot with a given figure height and width.
}

