% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_any.R
\name{grep_any}
\alias{grep_any}
\title{Multiple pattern matching}
\usage{
grep_any(x, y, match = "any", ignore.case = TRUE)
}
\arguments{
\item{x}{A vector of elements to search for}

\item{y}{A vector to search within}

\item{match}{Either 'any', 'exact' or 'begins'. Default is 'any'.}

\item{ignore.case}{Ignore case? Default is TRUE}
}
\description{
Apply \code{grep} to a pattern that can be comprised of multiple elements.
}
\examples{
x<-c(1,5,8,25)
y<-seq(0,20)
grep_any(x,y,'exact')
x<-c('4t','4r')
y<-c('3PN','4VN','4T')
grep_any(x,y)
}
