% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmap_dims}
\alias{heatmap_dims}
\title{Find the dimensions of a ggplot2 heatmap.}
\usage{
heatmap_dims(p)
}
\arguments{
\item{p}{a ggplot2 heatmap (or any ggplot with x and y mappings)}
}
\value{
a list with the number of rows and columns
}
\description{
This works by identifying the number of unique points in the variables mapped
to the x and y axes.
}
