% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{make_square}
\alias{make_square}
\title{Uses the number of columns and number of rows in a heatmap to fix the
aspect ratio such that the cells are perfectly square.}
\usage{
make_square(p, fudge = 1)
}
\arguments{
\item{p}{a ggplot object with discrete x and y aesthetics set}

\item{fudge}{a fudge factor multiplying the aspect ratio (< 1 = wider,
> 1 = taller). This does not appear to work right now.}
}
\value{
the same object, with the aspect ratio fixed to be the (number of
rows)/(number of columns)
}
\description{
Uses the number of columns and number of rows in a heatmap to fix the
aspect ratio such that the cells are perfectly square.
}

