% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{MakeHeatmapData}
\alias{MakeHeatmapData}
\title{Prepare agglomerated dataframe for heatmap display.}
\usage{
MakeHeatmapData(agg, s, min.rank.1 = "Genus", min.rank.2 = NA,
  min.samples = 1, hclust_method = "average", ...)
}
\arguments{
\item{agg}{an agglomerated df from \link{agglomerate}.}

\item{s}{sample metadata dataframe}

\item{min.rank.1}{the most specific rank to aggregate by (determines y-axis)
Uses \link{min_rank} to select next most specific rank if not available.}

\item{min.rank.2}{(optional) higher rank to prefix min.rank.1 (e.g. Phylum)
Uses \link{min_rank} to select next most specific rank if not available.}

\item{min.samples}{a row must appear more in than this number of samples to
be shown}

\item{hclust_method}{a character vector specifying the method to use for
clustering y-axis (from \link{hclust})}

\item{...}{additional samples passed to the min_rank calls}
}
\description{
Aggregates counts by minimum rank, completing missing cases, and clustering
rows. Expects input \code{agg} to have columns called \code{SampleID} and
\code{count} at minimum.
}
