% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{min_rank}
\alias{min_rank}
\title{Creates a MinRank vector from an agglomerated data frame.}
\usage{
min_rank(agg, end.rank, ranks = qiimer::taxonomic_ranks, ...)
}
\arguments{
\item{agg}{\link{agglomerate}d data frame (with taxonomic ranks)}

\item{end.rank}{the most specific taxonomic rank if available}

\item{...}{additional arguments passed to \link{tax_climber}}

\item{rank}{a character vector of taxonomic ranks present in agg}
}
\description{
Version of \link{tax_climber} for use inside \link{dplyr::mutate} or directly
assign to col. MinRank columns are the most specific taxonomic assignments
(up to a threshold)
}
