% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{Mode}
\alias{Mode}
\title{Find the mode of a logical or numeric vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a logical or numeric vector}
}
\value{
the mode
}
\description{
Find the mode of a logical or numeric vector
}
