% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{named_vector}
\alias{named_vector}
\title{Create a named vector using the values from one column in the second.}
\usage{
named_vector(.df, a, b)
}
\arguments{
\item{.df}{the dataframe to reference the columns from}

\item{a}{the column whose levels will be used as names}

\item{b}{the values of the named vector}
}
\value{
a named vector
}
\description{
The levels of `a` are used to pick the corresponding values in `b` so that
the named vector returned can be used in e.g. ggplot2::scale label
parameters.
}

