% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Plot a standard rainbow heatmap with faceting.}
\usage{
PlotHeatmap(heatmap.data, use.reads = TRUE, threshold = 0.6,
  x.text = TRUE, x.axis = "SampleID")
}
\arguments{
\item{heatmap.data}{the dataframe from \link{MakeHeatmapData}}

\item{use.reads}{if TRUE, plot the raw readcounts; if FALSE, use proportions.
Raw readcounts are subject to biases from uneven library size but can be
preferable for low-biomass or extremely sparse samples, as proportions can
be distorting.}

\item{threshold}{at what relative proportion the color scale goes to red.
This should be adjusted such that there is only one red cell per row.}

\item{x.text}{display labels on the x-axis?}

\item{x.axis}{name of factor to be used to define the x-axis (must exist in
\code{heatmap.data})}
}
\description{
Uses the output of \link{MakeHeatmapData} to display a heatmap with (by default)
SampleIDs on the x-axis and a minimum taxonomic rank of Genus on the y-axis.
}
