% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_tests.R
\name{polyatest}
\alias{polyatest}
\title{Uses a Dirichlet multinomial distribution to test for OTUs in the first
vector that are enriched.}
\usage{
polyatest(counts1, counts2, p.adjust.method = "fdr")
}
\arguments{
\item{counts1}{the first paired sample (and the target of the one-sided test)}

\item{counts2}{the second paired sample (i.e. prewash)}

\item{p.adjust.method}{the method to adjust p.values for multiple testing}
}
\value{
a vector of adjusted p values
}
\description{
Uses a Dirichlet multinomial distribution to test for OTUs in the first
vector that are enriched.
}

