% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny.R
\name{prune_tree}
\alias{prune_tree}
\title{Prunes a tree to only include specified OTUs.}
\usage{
prune_tree(tree, otus)
}
\arguments{
\item{tree}{an object of class "\code{phylo}"}

\item{otus}{a list of OTUs (or tip names) to keep}
}
\value{
pruned version of \code{tree}
}
\description{
Prunes a tree to only include specified OTUs.
}
