% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{read_blast6}
\alias{read_blast6}
\title{Reads an output file from BLAST in the outfmt=6 format (tab delimited)}
\usage{
read_blast6(fp, col.names = blast6.colnames)
}
\arguments{
\item{fp}{file path to blast output}

\item{col.names}{the column names of the BLAST output. The "std" columns are
defined in `blast6_colnames` (default)}
}
\description{
Reads an output file from BLAST in the outfmt=6 format (tab delimited)
}

