% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_duplicates.R
\name{rename_duplicates}
\alias{rename_duplicates}
\title{Rename duplicates}
\usage{
rename_duplicates(x, sep = "_", type = "alphabet", all = TRUE)
}
\arguments{
\item{x}{A vector}

\item{sep}{Specify the separator (delimiter) to use before the suffix}

\item{type}{Use letters (alphabet) or numbers to create the suffix}

\item{all}{Add the suffix to elements of the duplicated value, including the first instance (all=TRUE), or just to the subsequent instances}
}
\description{
Rename duplicates by adding a suffix
}
\examples{
x<-c(404, 405,406, 407, 408, 408 ,408, 408,409 ,409,410,411)
rename_duplicates(x)
rename_duplicates(x,all=FALSE)
rename_duplicates(x,type="numeric")
rename_duplicates(x,type="numeric",all=FALSE)
}
