% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{reorder_taxa}
\alias{reorder_taxa}
\title{Reorder levels of subtaxa by parent taxa}
\usage{
reorder_taxa(taxonomy, ranks = qiimer::taxonomic_ranks)
}
\arguments{
\item{taxonomy}{the taxonomy data frame, with columns for each [kpcofgs] rank}
}
\value{
the taxonomy with each column converted to a factor, and the levels
   reordered
}
\description{
Reorders the levels of each taxonomic rank to be arranged by the rank
   above it, so that e.g. each class level is ordered by the phylum it
   belongs to, and so on.
}

