% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_files}
\alias{search_files}
\title{Search for a pattern within text files}
\usage{
search_files(pattern, files, return.path = FALSE)
}
\arguments{
\item{pattern}{Case sensitive pattern to look for}

\item{files}{Text files to search (e.g., .R, .txt...)}

\item{return.path}{Include full file name path? Logical. Default is FALSE}
}
\description{
Returns names of files whose text contains a pattern
}
\examples{
#files<-list.files(path="...",full.names=TRUE, recursive = TRUE)
#files<-files[grep('.R',files)]
#search_files('some text',files)
}
\seealso{
readLines, grep
}
