% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_on_delimiter.R
\name{split_on_delimiter}
\alias{split_on_delimiter}
\title{Split data frame column or string on delimiter}
\usage{
split_on_delimiter(x, delimiter, col.names = NULL)
}
\arguments{
\item{x}{Data frame or string}

\item{delimiter}{Delimiter}

\item{col.names}{Names for the returned data frame columns}
}
\value{
A data frame
}
\description{
Split data frame column or string on delimiter
}
\examples{
string<-c("161293 - a" ,"155541 - b" , "161293 - c")
split_on_delimiter(string," - ")
}
