% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{subcolor_pal}
\alias{subcolor_pal}
\title{Grouped color palette for factors}
\usage{
subcolor_pal(fctr, h = c(0, 360) + 15, c = 100, l = 65, l.range = 20,
  h.start = 0, direction = 1)
}
\arguments{
\item{fctr}{the factor to create color scale for}

\item{h}{the range of hues to use}

\item{c}{chroma value to use}

\item{l}{the base luminance [0,100]}

\item{l.range}{the luminance above and below the base to use for subcolors}

\item{h.start}{the starting hue}

\item{direction}{1=clockwise, 0=counterclockwise around color wheel}
}
\value{
a named vector of hexadecimal colors
}
\description{
Picks a starting hue for each level of the factor, then returns colors
   with that hue, but varying luminosity for each item within that factor.
}

