% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{subset_matrix}
\alias{subset_matrix}
\title{Subset a (counts) matrix based on the column from a dataframe.}
\usage{
subset_matrix(s, mat, colname = "SampleID")
}
\arguments{
\item{s}{a sample data dataframe}

\item{mat}{a counts matrix}

\item{colname}{the name of the column to use for subsetting}
}
\value{
the subsetted matrix
}
\description{
This is useful for subsetting a counts matrix based on the sample IDs present
in the sample data dataframe. It then prunes the matrix to remove any rows
whose sums are 0.
}

