% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\name{tax_climber}
\alias{tax_climber}
\title{Finds the lowest taxonomic rank that isn't NA, starting at `start`.}
\usage{
tax_climber(otu, taxonomy, start = "Genus")
}
\arguments{
\item{otu}{the otu to look for}

\item{taxonomy}{the taxa annotations, with otus as rownames (important!)}

\item{start}{the starting taxonomic rank}
}
\value{
the lowest non-NA taxonomic rank assignment (with rank as name)
}
\description{
Finds the lowest taxonomic rank that isn't NA, starting at `start`.
}

