% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclectic.R
\docType{data}
\name{blast6_colnames}
\alias{blast6_colnames}
\title{Default BLAST outfmt 6 column names}
\format{An object of class \code{character} of length 12.}
\usage{
blast6_colnames
}
\description{
Default BLAST outfmt 6 column names
}
\keyword{datasets}
