% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcolors.R
\name{subcolor_pal}
\alias{subcolor_pal}
\title{Grouped color palette for factors}
\usage{
subcolor_pal(x, f, h = c(0, 360) + 15, c = 100, l = 65, l.range = 20,
  h.start = 0, direction = 1, na.value = "grey")
}
\arguments{
\item{x}{the vector or factor grouped by f (i.e. Genus)}

\item{f}{the vector or factor that groups the values of x (i.e. Order)}

\item{h}{the range of hues to use}

\item{c}{chroma value to use}

\item{l}{the base luminance [0,100]}

\item{l.range}{the luminance above and below the base to use for subcolors}

\item{h.start}{the starting hue}

\item{direction}{1=clockwise, 0=counterclockwise around color wheel}
}
\value{
a list containing `major`: the base color for each level in f; and
   `minor`: the color for each item in x
}
\description{
Picks a starting hue for each level of the factor, then returns colors
   with that hue, but varying luminosity for each item within that factor.
}
