% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_climber}
\alias{tax_climber}
\title{Finds the lowest taxonomic rank that isn't NA, stopping at `end`.}
\usage{
tax_climber(otus, taxonomy, end = "Genus", label = FALSE, sep = ":",
  ranks = qiimer::taxonomic_ranks)
}
\arguments{
\item{otus}{the otus to look for}

\item{taxonomy}{the taxa annotations, with otus as rownames (important!)}

\item{end}{the lowest desired taxonomic rank}

\item{label}{label the lowest rank with [kpcofgs]?}

\item{sep}{separator to use between rank label and rank value}

\item{ranks}{a character vector of taxonomic ranks, in order}
}
\value{
the lowest non-NA taxonomic rank assignment
}
\description{
Finds the lowest taxonomic rank that isn't NA, stopping at `end`.
}
